Implement support for response hijacking
When the response contains a `rack.hijack` header the response is handled asynchronously by handing off an IO like object to the callback in that header (after first committing the other response headers).

This is only half of the rack.hijack spec, and it might be the only part that is possible to support – the request part requires total control over the output, not even headers should be sent. Unfortunately there is no way to tell the application which part is supported so this sets the `rack.hijack?` property to true, and sets `rack.hijack` to a proc that raises NotImplementedError. It's not pretty, but at least it passes Rack::Lint.